+++ examples/rust_test/rust/src/nuttx_hal.rs
@@ -83,8 +83,8 @@ impl v2::InputPin for InputPin {
 
     /// Return true if GPIO Input is high
     fn is_high(&self) -> Result<bool, Self::Error> {
-        let mut invalue: i32 = 0;
-        let addr: *mut i32 = &mut invalue;
+        let mut invalue: isize = 0;
+        let addr: *mut isize = &mut invalue;
         let ret = unsafe {
             ioctl(self.fd, GPIOC_READ, addr)
         };
@@ -108,8 +108,8 @@ impl v2::InputPin for InterruptPin {
 
     /// Return true if GPIO Input is high
     fn is_high(&self) -> Result<bool, Self::Error> {
-        let mut invalue: i32 = 0;
-        let addr: *mut i32 = &mut invalue;
+        let mut invalue: isize = 0;
+        let addr: *mut isize = &mut invalue;
         let ret = unsafe {
             ioctl(self.fd, GPIOC_READ, addr)
         };
@@ -146,7 +146,7 @@ impl v2::OutputPin for UnusedPin {
 impl DelayUs<u32> for Delay {
     /// Sleep for us microseconds
     fn delay_us(&mut self, us: u32) {
-        unsafe { usleep(us); }
+        unsafe { usleep(us as usize); }
     }
 }
 
@@ -154,7 +154,7 @@ impl DelayUs<u32> for Delay {
 impl DelayMs<u32> for Delay {
     /// Sleep for ms milliseconds
     fn delay_ms(&mut self, ms: u32) {
-        unsafe { usleep(ms * 1000); }
+        unsafe { usleep(ms as usize * 1000); }
     }
 }
 
@@ -278,7 +278,7 @@ pub struct Spi {
     #[allow(dead_code)]
     path: *const u8,
     /// NuttX File Descriptor
-    fd:   i32,
+    fd:   isize,
 }
 
 /// NuttX GPIO Input
@@ -287,7 +287,7 @@ pub struct InputPin {
     #[allow(dead_code)]
     path: *const u8,
     /// NuttX File Descriptor
-    fd:   i32,
+    fd:   isize,
 }
 
 /// NuttX GPIO Output
@@ -296,7 +296,7 @@ pub struct OutputPin {
     #[allow(dead_code)]
     path: *const u8,
     /// NuttX File Descriptor
-    fd:   i32,
+    fd:   isize,
 }
 
 /// NuttX GPIO Interrupt
@@ -305,7 +305,7 @@ pub struct InterruptPin {
     #[allow(dead_code)]
     path: *const u8,
     /// NuttX File Descriptor
-    fd:   i32,
+    fd:   isize,
 }
 
 /// NuttX GPIO Unused
